// client/src/routes/constants/paths.ts
export enum paths {
    HOME = '/',
    CREATE_ARTICLE = '/create-article',
    ABOUT = '/about',
    BLOG_DETAILS = '/blog-details',
    BLOG = '/blog',
    CHANGELOG = '/changelog',
    CONTACT_US = '/contact-us',
    FAQ = '/faq',
    LICENCE = '/licence',
    REGISTRATION = '/registration',
    LOGIN = '/login',
    PRICING = '/pricing',
    PROJECT_DETAILS = '/project-details',
    PROJECT = '/project',
    SERVICE_SINGLE = '/service-single',
    SERVICES = '/services',
    TEAM = '/team',
    TEAM_SINGLE = '/team-single',
    PAGE_404 = '/page404',
}